+++ docs/changelog.md
@@ -30,6 +30,7 @@ The main goal of v0.6 was to simplify and stabalize the video input and statsfil
  * New `-b`/`--backend` option can be set to use a specific video backend
      * Current options are `opencv` (default) and `pyav` if the `av` package is installed
      * Run `scenedetect help` to see a list of backends available on the current system
+ * `-v`/`--verbosity` now takes precedence over `-q`/`--quiet`
  * `detect-content` command:
      * The default threshold value `-t`/`--threshold` has been lowered to 27 and is now slightly more sensitive to shot changes ([#246](https://github.com/Breakthrough/PySceneDetect/issues/246))
  * `detect-threshold` command: